Fixed a warning in Ordering.cpp.
Ordering was checking against a maximum queue size declared in the header which was a signed int. It became an unsigned long.

+++ Plugins/Ordering.h
@@ -29,7 +29,7 @@ private:
 	std::map<ConnectionID, uint32_t> nextExpectedMessage;
 	std::map<ConnectionID, uint32_t> nextOutgoingMessage;
 
-	const static int MAX_QUEUE_SIZE = 20;
+	const static unsigned long MAX_QUEUE_SIZE = 20;
 public:
 	Ordering();
